Move Clock to core
