+++ crates/eth2util/src/enr.rs
@@ -119,7 +119,7 @@ pub enum EnrEntry {
 impl EnrEntry {
     /// Writes this entry into the record's key-value pairs.
     fn apply(self, kvs: &mut HashMap<String, Vec<u8>>) {
-        match self {
+        let _ = match self {
             Self::Ipv4(ip) => kvs.insert(KEY_IP.to_string(), ip.octets().to_vec()),
             Self::Tcp(tcp) => kvs.insert(KEY_TCP.to_string(), tcp.to_be_bytes().to_vec()),
             Self::Udp(udp) => kvs.insert(KEY_UDP.to_string(), udp.to_be_bytes().to_vec()),
